🚚 api: move cpf service to src/shared context
