+++ api/tests/certification/session-management/unit/infrastructure/serializers/certification-serializer_test.js
@@ -1,7 +1,7 @@
-import { WrongDateFormatError } from '../../../../../../src/shared/domain/errors.js';
 import * as serializer from '../../../../../../src/certification/session-management/infrastructure/serializers/certification-serializer.js';
 import { CertificationCourse } from '../../../../../../src/certification/shared/domain/models/CertificationCourse.js';
 import { NO_EXAMINER_COMMENT } from '../../../../../../src/certification/shared/domain/models/CertificationReport.js';
+import { WrongDateFormatError } from '../../../../../../src/shared/domain/errors.js';
 import { domainBuilder, EMPTY_BLANK_AND_NULL, expect } from '../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | certification-serializer', function () {
